+++ Modules/booking/View/Booking/bookweek.php
@@ -13,11 +13,15 @@ $toDate = $sundayDate;
 $nbDays = 7;
 $resourcesBase = $resourceBase ? [$resourceBase]: [];
 $calEntries = [$calEntries];
-$from = ["week", $date, $bk_id_resource, $bk_id_area, $id_user];
+$from = ["week", $date, $bk_id_resource, $bk_id_area, $id_user, $detailedView ? 'detailed' : 'simple'];
 $isUserAuthorizedToBook = [ $isUserAuthorizedToBook];
 
-echo drawNavigation('week', $id_space, $startDate, $toDate, $beforeDate, $afterDate, $bk_id_resource, $bk_id_area, $id_user, $lang);
-include 'Modules/booking/View/Booking/caldisplay.php';
+echo drawNavigation('week', $id_space, $startDate, $toDate, $beforeDate, $afterDate, $bk_id_resource, $bk_id_area, $id_user, $detailedView, $lang);
+if($detailedView) {
+    include 'Modules/booking/View/Booking/caldisplay.php';
+} else {
+    include 'Modules/booking/View/Booking/simplecaldisplay.php';
+}
 		
 ?>