+++ CHANGELOG.md
@@ -5,11 +5,13 @@
 * [core] add redirects to TODO after doing todo actions
 * [core] use bootstrap5 and update/remove some libraries
 * [core] show number of user clients in user admin view
+* [booking] add to calendar a summary view option
+* [core][ldap] do not automatically add ldap users to spaces
 
 ## 2.4.1
 
 * [core] add missing tableName properties in model constructors
-* [space] improve todolist sql resquests
+* [space] improve todolist sql requests
 * [helpdesk] close connection in anycase
 * [core] fix TODO in corespaceadmin, do not get all bookings, just count them
 * [core] if module is inactive, fix authorization (isUserMenuSpaceAuthorized)